requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
backcall==0.2.0
debugpy==1.3.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
decorator==5.0.9; python_version >= '3.5'
greenlet==1.1.0; python_version >= '3'
ipykernel==6.0.3
ipython-genutils==0.2.0
ipython-sql==0.4.0
ipython==7.25.0; python_version >= '3.7'
jedi==0.18.0; python_version >= '3.6'
jupyter-client==6.1.12; python_version >= '3.5'
jupyter-core==4.7.1; python_version >= '3.6'
matplotlib-inline==0.1.2; python_version >= '3.5'
numpy==1.21.1; python_version >= '3.7'
pandas==1.3.0
parso==0.8.2; python_version >= '3.6'
pexpect==4.8.0; sys_platform != 'win32'
pickleshare==0.7.5
prettytable==0.7.2
prompt-toolkit==3.0.19; python_full_version >= '3.6.1'
psycopg2==2.9.1
ptyprocess==0.7.0
pygments==2.9.0; python_version >= '3.5'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==0.18.0
pytz==2021.1
pyzmq==22.1.0; python_version >= '3.6'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlalchemy==1.4.21; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
sqlparse==0.4.1; python_version >= '3.5'
tornado==6.1; python_version >= '3.5'
traitlets==5.0.5; python_version >= '3.7'
wcwidth==0.2.5
